+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## 0.37.8
+
+### Patch Changes
+
+- 80ae089: - **feat:**: Enhanced `Identity` hooks with full TanStack Query options support. By @cpcramer #2116
+  - **feat:**: Added MiniKit Provider and Hooks. By @alessey #2082, #2101, #2112
+  - **fix:**: Fixed WalletDropdownLinkReact type to allow ReactNode in icon property. By @gmolki #2058
+  - **feat**: Added event handlers to `Earn` components. @abcrane123 #2106
+
+## 0.37.7
+
+### Patch Changes
+
+- **feat**: Batch support for Basenames and ENS names. By @cpcramer #2102
+- **feat**: Added Batch support for Base and Ethereum Avatars. By @cpcramer #2107
+- **feat**: Added base mainnet deploychain address. By @0xAlec #2096
+- **feat**: Implemented Swap default children. By @alessey #2077
+- **fix**: Improved Next.js server env compatibility. By @dschlabach #2090
+
 ## 0.37.6
 
 ### Patch Changes